refactor: rename NowIndicator's file to match its name
