+++ lib/src/components/now_indicator.dart
