+++ lib/timetable.dart
@@ -10,7 +10,7 @@ export 'src/components/month_widget.dart';
 export 'src/components/multi_date_content.dart';
 export 'src/components/multi_date_event_header.dart';
 export 'src/components/multi_date_header.dart';
-export 'src/components/now_indicator_painter.dart';
+export 'src/components/now_indicator.dart';
 export 'src/components/time_indicator.dart';
 export 'src/components/time_indicators.dart';
 export 'src/components/week_indicator.dart';
